+++ encoding/test.ts
@@ -1,2 +1,2 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-import "./parser_test.ts";
+import "./toml_test.ts";
