+++ test.ts
@@ -3,7 +3,7 @@
 import "./archive/tar_test.ts";
 import "./colors/test.ts";
 import "./datetime/test.ts";
-import "./encoding/toml/test.ts";
+import "./encoding/test.ts";
 import "./examples/test.ts";
 import "./flags/test.ts";
 import "./fs/test.ts";
